Emailing_Service-master/requirements.txt
apscheduler==3.6.3
boto3
django==3.0.8
djangorestframework==3.11.1
django_amazon_ses==3.0.2
dnspython==2.0.0
requests==2.24.0
mysqlclient